\name{edf-package}
\alias{edf-package}
\alias{edf}
\docType{package}
\title{
EDF/EDF+
}
\description{
Import EDF/EDF+ files.
}
\details{
\tabular{ll}{
Package: \tab edf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-03\cr
License: \tab GPL3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ryan M. Hope

Maintainer: Ryan M. Hope <rmh3093@gmail.com>
}
\keyword{ package }